backend/requirements.txt
Flask[async]
python-dotenv
openai
urllib3==1.26.18
APScheduler
